Rename nihms-model to deposit-model
With the decision to use a single model for all deposit service targets,
it is appropriate to rename the existing model.  The model had been
created when initial development only targeted a NIHMS server, and so
that name had been used.  Soon, the model will also be used when
targeting a DSpace server, and so a more generic name is now preferable.

This commit renames the model project and its classes and updates
dependencies and references throughout the nihms-submission module.
Later renaming efforts will address other projects and ultimately change
the package groupId.

A few additional edits are included, and several references to NIHMS
were let in place until their appropriateness can be further evaluated.