Builder modules renamed to be target-agnostic
Since the deposit service data model will be used by all targets, the
modules that define how the models are built should have target-agnostic
names.  Here, nihms-builder is renamed to builder-api (to match the
naming for the assembler and transport api modules) and
nihms-filesystem-builder is renamed to filesystem-builder.  Shortly, a
new fcrepo-builder module will be added to build models from data stored
in a Fedora repository.