Alter a previous renaming of what was nihms-filesystem-builder
Originally, nihms-filesystem-builder was renamed to filesystem-builder,
but as I was working to reconcile that with the need for very similar
functionality to read from a Fedora repository, it became clear that
instead of having three projects (one to copy data and two derived from
it for filesystem and Fedora data), that this should all be done in one
module with three classes for the separate functionality.  Then, other
modules derived from builder-api would be for data that came from places
other than Fedora.