Initial implemetation to import Fedora into Deposit Service
This submission contains the initial implementation of several classes
that support reading PASS Fedora data into the deposit service data
model.  PASS data is accessed through the pass-client PassEntity data
model, which is generated either by reading data from a Fedora
repository using a FedoraPassClient or deserialized from a JSON file
using a PassJsonAdapter.

Both methods of acquiring Fedora data produce a list of PassEntity
objects and identify the Submission entity at the root of the tree of
resources that  fully describe the data needed by the deposit service.
The resources in the list may only reference other resources in the
list, and the list must contain exactly one Submission resource.

ModelBuilder is an abstract base class that performs the data conversion
from PassEntity to the deposit-model's DepositSubmission.  In this
commit, the code walks the entirety of the PassEntity tree and copies
some properties to the DepositSubmission structure.  Because both the
Fedora data model and the deposit data model are still in flux, there
are many comments in the code identifying places where data from the
PassEntity objects is not used and where values in the DepositSubmission
are not being filled.  These should be resolved as model decisions are
made.

FilesystemModelBuilder was revised to read a JSON file full of
serialized PassEntity data instead of using a file of Java properties.
FcrepoModelBuilder was added to read PassEntity data from a Fedora
repository.  Both rely on the new PassJsonFedoraAdapter, which can
serialize/deserialize between a list of PassEntity objects and JSON
text, and can upload/download between a list of PassEntity objects  and
a Fedora server.  We may want to eventually move this class to
pass-client for greater accessibility.

Unit tests for all three of these classes rely at present on a single
JSON sample data file that was added to the project.  This data is
completely made up and needs to be formalized as required by data model
changes and the other deposit service modules.