Fix NihmsSubmissionAppFtpIT to work with fcrepo-builder
The NihmsSubmissionAppFtpIT test was not updated to work with recent
changes to the deposit service "builder" classes, and so it was failing
on Travis.  Here, the test is changed to work with JSON sample data
representing PassEntity objects rather than a set of name/value
properties, much as was done for the fcrepo-builder unit tests.

The deposit service integration test module was also (through some
dependencies) loading an older version of the joda-time artifact, which
caused PassEntity parsing to fail.  An exclusion was added to the POM to
prevent this dependency from being resolved with old artifact version.

Some small changes were performed in other classes to capitalize some
static string variable names and satisfy StyleCheck.  The
Submission.name property is also initialized in the existing deposit
service model, despite the absence of a true name value, as it cannot be
null.