Change fedora-builder unit tests to be integration tests
The previously created (but inactive) unit tests in the fedora-builder
module are moved to the nihms-integration module.  These tests were
confirmed to work against a Fedora server running from the configuration
in nihms-integration, but still need to be tested without a previously
started server to ensure the server runs as expected.