Add Fedora server to integration suite and include fedora-related ITs (#47)
* Adds Fedora server to nihms-integration Docker containers.

* Upgrade fedora-builder project to use latest client data model

After the iniitial implementation of the fedora-builder module, the
fedora client's data model underwent substantial changes.  This
submission brings fedora-builder up to date with those changes.

The unit tests included here pass when running against a Fedora server
that supports the new data model, but have been commented out until a
Fedora server can be included in this project's integration testing
suite.

* Upgrade fedora-builder project to use latest client data model

After the iniitial implementation of the fedora-builder module, the
fedora client's data model underwent substantial changes.  This
submission brings fedora-builder up to date with those changes.

The unit tests included here pass when running against a Fedora server
that supports the new data model, but have been commented out until a
Fedora server can be included in this project's integration testing
suite.

* Update sample data file in IT module

This file had been copied from sample data in a different project, and a
recent submission updated that content to match the latest data model
changes.

* Change fedora-builder unit tests to be integration tests

The previously created (but inactive) unit tests in the fedora-builder
module are moved to the nihms-integration module.  These tests were
confirmed to work against a Fedora server running from the configuration
in nihms-integration, but still need to be tested without a previously
started server to ensure the server runs as expected.

* Adds Fedora server to nihms-integration Docker containers.

* Change fedora-builder unit tests to be integration tests

The previously created (but inactive) unit tests in the fedora-builder
module are moved to the nihms-integration module.  These tests were
confirmed to work against a Fedora server running from the configuration
in nihms-integration, but still need to be tested without a previously
started server to ensure the server runs as expected.