Find Files related to a Submission using Incoming Links
The previous implementation to find Files that reference a Submission
used the indexer to get a list of such resources.  Problems with the
indexer made this implementation unreliable, so it has been changed to
use new functionality in the Fedora Client that returns a list of
resources that link to a specified resource.  This list is filtered to
only include File resources.

Also included is some cleanup to of the ITs that invoke this
functionality so that Fedora resources created for the tests are removed
from the server after the tests finish.  A third File resource was
removed from the sample data to simplify the tests.  A few POM
dependencies were improved.