A fresh start on a PR for binary files that reference a submission (#53)
All of the File resources that reference the supplied Submission
resource will be retrieved from the Fedora server and included in the
deposit service data model that is being constructed.  Sample data is
updated to contain three File resources, only two of which reference the
data's Submission.

* Find Files related to a Submission using Incoming Links

The previous implementation to find Files that reference a Submission
used the indexer to get a list of such resources.  Problems with the
indexer made this implementation unreliable, so it has been changed to
use new functionality in the Fedora Client that returns a list of
resources that link to a specified resource.  This list is filtered to
only include File resources.

Also included is some cleanup to of the ITs that invoke this
functionality so that Fedora resources created for the tests are removed
from the server after the tests finish.  A third File resource was
removed from the sample data to simplify the tests.  A few POM
dependencies were improved.

* Use a specific exception type when finding Files that reference a Submission.

* Make the exception handling during File-finding more robust

Search the root causes of the exception for one that is specific to the
type conversion used in the fedora client.  If no such exception is
found, print the original exception.

* Throw an unacceptable exception rather than print it

Other types of unacceptable exceptions in this method do not have their
stacks printed, so dealing with any of these exceptions is the
responsibility of the caller.