Parse a Submission's metadata
The Submission object that is to be deposited in target repositories
contains a JSON string specifying some metadata that may be essential
for performing the deposit.  These changes parse that data and where
appropriate, use the properties it contains to initialize the deposit.
Further refinement will be needed when the deposit service metadata
structures are updated to meet the needs of both the NIHMs and DSpace
targets.