Better processing of Embargo metadata (#57)
* Better processing of embargo metadata

A submission's metadata's embargo properties are optional, so the
parsing code now keys on the presence of the "under-embargo" property
before trying to parse any of the properties.  Values of all properties
are now stored in new members of DepositMetadata's Article class.
Parsing of the embargo end date (as it currently exists in the sample
data set) was updated to allow the value to be stored as a
ZonedDateTime.  This may have to be revisited if the front end produces
more complicated date-time values or uses a different date format.