Clean up deposit service data model and NIHMS export (#79)
* Deposit model and export cleanup

The PubMed-related members of the deposit service data model are removed
and no longer exported.  The DOI that is exported is now always in "raw"
form, with no leading URI scheme or host.

* Modify some DOI data and tests

For data and tests that are NIHMS-only, modify the data or comparison
strings to account for removal of URI scheme and host.