Fixes for PI/author data in NIHMS metadata (#91)
* Fixes for PI/author data in NIHMS metadata

The PI info in the NIHMS metadata file should be based on the
information in a Submission resource.  There should only be one PI
listed here, and they should be noted as the corresponding PI.  The
Submission does not include any information about whether this PI is an
author, so that property is now omitted.

Some naming confusion is also cleared up regarding the "copi" members of
a Grant.  They are now added to the deposit metadata people list, but no
longer as corresponding PIs.  None of these PIs are flagged as authors,
as that information is not present at the time of their creation.

* Update tests for author/PI changes

* Strengthen unit test.