Fixed bug where one Assembler's modification of the DepositSubmission cause another Assembler to fail.
The DepositSubmission model is now shared between all Assemblers (see SubmissionProcessor).

NihmsAssembler modifies the DepositManifest by adding a DepositFile for the bulk submission metadata.  The bulk submission metadata doesn't actually exist anywhere, it is generated by the NihmsMetadataSerializer, and as such, the DepositFile representing the metadata has no location URI.

By the time NihmsAssembler modifies the DepositManifest, the DepositManifest has already been "validated", insuring that each DepositFile in the manifest has a valid location URI.  Therefore, assembly of the NIHMs package proceeds without error.

If another Assembler is executed over the same DepositSubmission, it will re-validate the DepositManifest.  First, the DepositManifest contains a DepositFile (the bulk submission metadata added by NihmsAssembler) that it shouldn't.  Second, the DepositFile has no location URI, so when the second Assembler attempts to resolve the bytes associated with the DepositFile, an NPE is thrown when resolving the content.

This hack simply updates the NimsManifestSerializer to include a line for the NIH bulk submission metadata if it is not already included in the DepositManifest.