Revise deposit service metata handling of submittors, PIs and authors
Previously, the deposit service metadata for "people" included boolean flags to indicate whether an individual was a PI, a coPI or an author.  This system failed to denote which people were submitters, who must be included in NIHMs deposit metadata but not in DSpace contributor lists.  This submission replaces the set of boolean flags with an enumeration listing the four types of people.

The ModelBuilder was modified to assign types to people it was adding to the deposit service data model.  Authors come from the Submission metadata, and they were previously being ignored.  Adding them required allowing for person names that are a single string rather than the first/middle/last name strings provided by the User resources that provide the submitter and PIs.  Now, wherever possible the Person object is asked to provide the total name, which is either the single string name (if provided) or a constructed name from the other three fields.

Note that a single human can be added to the Person list more than once, with each entry corresponding to a single role.  Multiple entries were already possible due to the three People sources.  The three sources and different name formats make it difficult to find sets of People who represent the same human.  Unless this is resolved, the same human may be listed more than once in the DSpace contributors list.

Due to the Person matching problem mentioned above, the NIHMs metadata was changed slightly to not include the "pi" attribute for the submitter (corresponding PI).
 The attribute is optional and so can be removed.  Submitters are no longer included in the DSpace contributor list, but authors are now included.

Several unit tests were updated to check the validity of the data model that is created from Fedora and the metadata that is assembled for the deposits.