Initial implementation for issue #147
These changes implement the inclusion of additional metadata for JScholarship deposits as specified in card #147 (including the comments).  The implementation is complete except for the inclusion of a publication date in the citation, as that value may not be available in our submission metadata.

An open question is what the XML element name should be for recognition by JScholarship.  The existing implementation (for a generic DSPACE target) uses an element name that includes "qualified", which is not practially true of this JScholarship implementation.

Additional unit and integration tests will be added to this work in upcoming submissions.