Initial implementation of  enhanced metadata (#151)
* Initial implementation for issue #147

These changes implement the inclusion of additional metadata for JScholarship deposits as specified in card #147 (including the comments).  The implementation is complete except for the inclusion of a publication date in the citation, as that value may not be available in our submission metadata.

An open question is what the XML element name should be for recognition by JScholarship.  The existing implementation (for a generic DSPACE target) uses an element name that includes "qualified", which is not practially true of this JScholarship implementation.

Additional unit and integration tests will be added to this work in upcoming submissions.

* Fix and augment unit tests related to DSPACE metadata

These changes fix some failing unit tests, augment others and add a new test for the DSPACE metadata.  They also limit the contributors who are included in the citation to only be those who are listed as authors.

* Handle publication data from submission metadata

The "publicationDate" element of the submission metadata is now parsed by the ModelBuilder and saved in the deposit services data model.  From there, it is included in the citation that is produced for JScholarship.  Tests were updated to include the publication date.

* Revise format of abstract and citation in metadata

Because the previous changes produced metadata that was not accepted, this submission revises the way the abstract and citation data is created.

* Handle metadata with contributors specified only by full name

The previous implementation for generating a citation assumed that contributor metadata always had values for the first/middle/last names.  This is not the case, and this change used the full-name value if the other properties are not set.