Print environment variables and Spring property resolutions to deposit service log.
When the deposit service starts it will print an alphabetized list of all defined environment variables and their values.  This will be followed by a list of all Spring property definitions and their resolved values.  If an error occurs during property resolution, it is printed instead of a resolved value.