Fix for broken unit test
The previous commit contained a change that saves a contributor's display name to the deposit services data model.  This unexpectedly broke a unit test in NIHMS due to that display name being used for a comparison against a name that was built up from first/middle/last names.  Here, a new method is introduced in the data model to explicitly get a name build up from first/middle/last names and that method is used to fix the problem.

After this fix, the ITs run on my system up until SubmissionProcessorIT, which fails (as we've seen happen with a recent submission to master).  That test passes on my system when run in isolation, so I expect this is related to the known problem.