Merge metadata PR from maintenance branch into master (#176)
* Fix DSpace metadata issues and add unit tests for metadata

Deposit Services issue 165 points out a couple of problems with the way submission metadata is presented in DSpace deposit metadata.  This change fixes those issues and adds two unit tests that cover the path of metadata from Submission resource through the deposit services data model to the DSpace XML metadata.

Specific fixes are to:
* Make sure the publisher name is included in the output when it is specified in either the "common" or "crossref" sections of the metadata, with crossref taking precedence.
* Use the publication date from either the common or crossref metadata sections when creating the output metadata's citation entry.
* Use a single "display" name value for PI contributor entries when the first/middle/last names are not provided.
* Extend an existing unit test to include data like that which was causing problems.

The two new tests use two new sample data sets (JSON representations of Submission resources with metadata).  The first test tests the cases of:
* Publisher name and publication date being specified only in the common metadata section
* Submitters not generating contributor entries.
* Citation format for three or fewer authors.

The second test covers the cases of:
* Publisher name and publication date specified in the crossref metadata section.
* Citation format for four or more authors.

* Fix for broken unit test

The previous commit contained a change that saves a contributor's display name to the deposit services data model.  This unexpectedly broke a unit test in NIHMS due to that display name being used for a comparison against a name that was built up from first/middle/last names.  Here, a new method is introduced in the data model to explicitly get a name build up from first/middle/last names and that method is used to fix the problem.

After this fix, the ITs run on my system up until SubmissionProcessorIT, which fails (as we've seen happen with a recent submission to master).  That test passes on my system when run in isolation, so I expect this is related to the known problem.

* Merge metadata commits with changes in master

Master included some changes that conflicted with the commits that were cherry picked from maintenance, and this commit reconciles them.