Deposit status API update (#180)
* Update the DepositStatus[Processor|Resolver] methods to accept the more general RepositoryConfig to be more stable with respect to future configuration changes.

* Decouple the RepositoriesFactoryBeanConfig from the DepositProcessing interface.

During a refactoring exercise (an effort to distill and separate a configuration module and a deposit status api) a cycle in the dependency graph appeared.  This caused problems not only from the perspective of separating out the code into separate Maven modules, it also manifested a cycle when building the Spring ApplicationContext.

This change allows the Repositories object to be built, and lazily look up the DepositProcessor when building Packagers, rather than looking up the DepositProcessor while the Repositories are being built.