Improvments to startup of Deposit Services - Import cleanup and removal of unused methods - Log the git build time on start - Protect against NPEs - Improved logging of the Packagers used - Improve logginging in JmsConfig when a Submission fails
Protect against NPEs when a configuration does not have a "deposit-processing" or "deposit-config" object.

Better reporting on Packager configuration during startup.

Add the build date to the Git information printed to the console on Deposit Services startup.

Improve logging when processing a JMS message fails.  There's a lot that happens when the submissionConsumer processes the Submission so the error message should be more generic.