Generify the SubmissionProcessorIT to work with the expected results from the FilesystemTransport, and be independent of any specific Repository in use.
The behavior of Deposit Services, in particular the results verified by SubmissionProcessorIT, are dependant on the Transport and Repositories used.  This update removes the dependence on specific Repositories by searching for Deposits using relationships specific to the Submission.

However, this test verifies the expected state of resources after a successful submission completes, and the state of those resources is dependent on the Transport used. This has always been true in Deposit Services from day one, but it has been made explict with the recent addition of the TransportResult.onSuccess(...) method.

At the moment, this IT as configured, will not test the processing of the Deposit.depositStatusRef field, because the FilesystemTransport does not use it.  This means that the processing of Deposit.depositStatusRef will need to be integration tested elsewhere.