Support for Submission.metadata collection hints (#269)
* Update SwordV2Binding to support 'collection-hints' configuration element.

* WIP: Update PackageStream.Metadata to carry PASS Submission.metadata.

Stage is set to update the model builder implementations, followed by:
- Updating the AbstractAssembler.assemble
- AssemblerSupport.buildMetadata
- ...

* WIP: Update PackageStream.Metadata to carry PASS Submission.metadata.

Update of the model builder to serialize the Submission.metadata as a JsonObject on the DepositSubmission.  Adds the Google GSON library as a dependency to the deposit-model.

* WIP: Update PackageStream.Metadata to carry PASS Submission.metadata.

Update of the model builder to serialize the Submission.metadata as a JsonObject on the DepositSubmission.  Adds the Google GSON library as a dependency to the deposit-model.  Updated FcrepoModelBuilderIT to test contents of DepositSubmission metadata JsonObject.

* WIP: Update PackageStream.Metadata to carry PASS Submission.metadata.

Update MetadataBuilder and PackageStream interfaces to represent the Submission.meta as a JsonObject (easiest) instead of a Map<String, Object>.  Updated impls (MetadataBuilderImpl, PreassembledAssembler, SimpleMetadataImpl).

Update AbstractAssembler to add the DepositSubmission.submissionMeta to the builder.

Update Sword2TransportSession to obtain hints from the submission meta carried by the PackageStream.Metadata.

* Update Sword2TransportSession to use the hints to select a collection.  The first configured hint that matches is used.  Any subsequent matches are ignored and do not affect the collection used for deposit.

* Fix travis.

* Bump version to 1.0.1-3.4-SNAPSHOT from 1.0.0-3.4.